requirements.txt
langchain-community
pgvector
pandas
python-dotenv
groq  
psycopg2-binary  
langchain-groq